+++ lib/types/Form.d.ts
@@ -1,0 +1,50 @@
+import { Access, ProjectId, SubmissionId } from '@formio/types';
+export type FormId = string;
+export interface Form {
+    _id: FormId;
+    _vid: number;
+    title: string;
+    name: string;
+    path: string;
+    type: FormType;
+    display?: FormDisplay;
+    action?: string;
+    tags?: string[];
+    access: Access[];
+    submissionAccess: Access[];
+    fieldMatchAccess: any;
+    owner: SubmissionId;
+    machineName: string;
+    components: Array<any>;
+    settings?: FormSettings;
+    properties: Record<string, string>;
+    project: ProjectId;
+    revisions: 'current' | 'original' | '';
+    submissionRevisions: 'true' | '';
+    controller?: string;
+    builder?: boolean;
+    page?: number;
+    created: Date | string;
+    modified: Date | string;
+    deleted: Date | string;
+}
+export type FormType = 'form' | 'resource';
+export type FormDisplay = 'form' | 'pdf' | 'wizard';
+export type FormSettings = {
+    collection?: string;
+    condensedMode?: boolean;
+    disableAutocomplete?: boolean;
+    fontSize?: number;
+    hideTitle: boolean;
+    layout?: string;
+    margins?: string;
+    showCheckboxBacground?: boolean;
+    theme?: string;
+    viewAsHtml?: boolean;
+    viewer?: string;
+    wizardHeaderType?: string;
+    pdf?: {
+        src: string;
+        id: string;
+    };
+};