+++ lib/types/project/settings/authorization/index.d.ts
@@ -1,0 +1,3 @@
+export * from './oauth';
+export * from './ldap';
+export * from './saml';