+++ lib/types/project/settings/authorization/ldap.d.ts
@@ -1,0 +1,7 @@
+export type ProjectLdapConfig = {
+    bindCredentials: string;
+    bindDn: string;
+    searchBase: string;
+    searchFilter: string;
+    url: string;
+};