+++ lib/utils/jwtDecode.d.ts
@@ -1,0 +1,3 @@
+export declare function jwtDecode(token: string, options?: {
+    header?: boolean;
+}): any;